Закончил с новым анализатором звука, вернул Праат пока не придумаю, как ускорить его

+++ Tuto/Model/Current/Montage/SoundInterval.cs
@@ -22,7 +22,7 @@ namespace Tuto.Model
         public bool HasVoice { get; set; }
 
         [DataMember]
-        public double Volume { get; set; }
+        public int Volume { get; set; }
         
         public int MiddleTime { get { return (StartTime + EndTime) / 2; } }
